requirements.txt
asn1crypto==0.24.0
backports.functools-lru-cache==1.5
beautifulsoup4==4.8.0
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
cryptography==2.1.4
Django==1.11.20
enum34==1.1.6
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
httplib2==0.13.1
idna==2.8
ipaddress==1.0.17
keyring==10.6.0
keyrings.alt==3.0
oauthlib==3.1.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycrypto==2.6.1
pygobject==3.26.1
pytz==2019.1
pyxdg==0.25
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
SecretStorage==2.3.1
six==1.12.0
soupsieve==1.9.2
uritemplate==3.0.0
urllib3==1.25.3
